Projet_Traduction_IA/requirements.txt
yt-dlp
openai-whisper
transformers
torch
sentencepiece
sacremoses
colorama
customtkinter